+++ CHANGELOG.md
@@ -19,12 +19,22 @@ Don't forget to remove deprecated code on each major release!
 
 ## [Unreleased]
 
+### Added
+
+-   Support for custom routers.
+
 ### Changed
 
 -   Set maximum ReactPy version to `<2.0.0`.
 -   Set minimum ReactPy version to `1.1.0`.
 -   `link` element now calculates URL changes using the client.
 -   Refactoring related to `reactpy>=1.1.0` changes.
+-   Changed ReactPy-Router's method of waiting for the initial URL to be deterministic.
+-   Rename `StarletteResolver` to `ReactPyResolver`.
+
+### Removed
+
+-   `StarletteResolver` is removed in favor of `ReactPyResolver`.
 
 ### Fixed